requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.22.0
drf-spectacular-sidecar==2022.4.1
inflection==0.5.1
jsonschema==4.4.0
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
